+++ client/lib/ft4/authentication/types.ts
@@ -12,7 +12,7 @@ export interface Authenticator {
   createSession(): AuthenticatorSession;
   getAuthFlags(operation: Operation): Promise<string[]>;
   getKeyHandlerForOperation(
-    operation: Operation
+    operation: Operation,
   ): Promise<KeyHandler | undefined>;
   getNonce(authDescriptorId: BufferId): Promise<number | null>;
 }
@@ -27,7 +27,7 @@ export interface KeyHandler {
     accountId: BufferId,
     operation: Operation,
     nonce: number,
-    authDataService: AuthDataService
+    authDataService: AuthDataService,
   ): Promise<Operation[]>;
 
   sign(transaction: TxBuilderTransaction): Promise<void>;
@@ -57,7 +57,7 @@ export interface AuthDataService {
   getAuthMessageTemplate(operation: Operation): Promise<string>;
   getNonce(
     accountId: BufferId,
-    authDescriptorId: BufferId
+    authDescriptorId: BufferId,
   ): Promise<number | null>;
   getLoginConfig(name: string | null): Promise<LoginConfig>;
 }