+++ client/lib/ft4/utils/transaction-builder.ts
@@ -34,7 +34,7 @@ export type TransactionBuilder = {
    */
   addWithAuthenticator: (
     operation: Operation,
-    authenticator: Authenticator
+    authenticator: Authenticator,
   ) => TransactionBuilder;
   /**
    * Add key handlers that will also be included as signers to this operation.
@@ -88,7 +88,7 @@ export type TransactionBuilder = {
  */
 export function transactionBuilder(
   authenticator: Authenticator,
-  client: IClient
+  client: IClient,
 ): TransactionBuilder {
   function add(operation: Operation): TransactionBuilder {
     this._operations.push([operation, authenticator]);
@@ -104,7 +104,7 @@ export function transactionBuilder(
 
   async function buildUnsigned() {
     const [operations, keyHandlers] = await authenticateOperations(
-      this._operations
+      this._operations,
     );
     keyHandlers.forEach((kh) => this._keyhandlersUsed.push(kh));
     const txn: TxBuilderTransaction = {
@@ -123,7 +123,7 @@ export function transactionBuilder(
   }
 
   async function authenticateOperations(
-    operations: OpAuthPair[]
+    operations: OpAuthPair[],
   ): Promise<[Operation[], KeyHandler[]]> {
     const keyHandlers: KeyHandler[] = [];
     const nonces = new Map<Buffer, number>();
@@ -134,11 +134,11 @@ export function transactionBuilder(
 
       if (
         !(await authenticator.authDataService.isOperationExposed(
-          operation.name
+          operation.name,
         ))
       ) {
         throw new OperationNotExistError(
-          `Operation ${operation.name} does not exist`
+          `Operation ${operation.name} does not exist`,
         );
       }
 
@@ -148,19 +148,19 @@ export function transactionBuilder(
       }
 
       const keyHandler = await authenticator.getKeyHandlerForOperation(
-        operation
+        operation,
       );
 
       if (!keyHandler) {
         throw new AuthorizationError(
-          `No keyhandler registered to handle operation <${operation.name}>`
+          `No keyhandler registered to handle operation <${operation.name}>`,
         );
       }
       keyHandlers.push(keyHandler);
       if (!nonces.has(keyHandler.authDescriptor.id)) {
         nonces.set(
           keyHandler.authDescriptor.id,
-          (await authenticator.getNonce(keyHandler.authDescriptor.id))!
+          (await authenticator.getNonce(keyHandler.authDescriptor.id))!,
         );
       }
 
@@ -169,7 +169,7 @@ export function transactionBuilder(
         authenticator.accountId,
         operation,
         nonce,
-        authenticator.authDataService
+        authenticator.authDataService,
       );
       // consider keeping nonce value in corresponding key handler
       ops.forEach((op) => {
@@ -191,7 +191,7 @@ export function transactionBuilder(
   async function build() {
     const tx = await this.buildUnsigned();
     await Promise.all(
-      this._keyhandlersUsed.map((handler: KeyHandler) => handler.sign(tx))
+      this._keyhandlersUsed.map((handler: KeyHandler) => handler.sign(tx)),
     );
     return gtx.serialize(tx);
   }
@@ -212,7 +212,7 @@ export function transactionBuilder(
 
   function addWithAuthenticator(
     operation: Operation,
-    authenticator: Authenticator
+    authenticator: Authenticator,
   ): TransactionBuilder {
     this._operations.push([operation, authenticator]);
     return this;